requirements.txt
fastapi
uvicorn
google-auth
requests
pydantic